+++ lib/editor/atto/plugins/cloze/lib.php
@@ -22,8 +22,6 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
-
 /**
  * Get the list of strings for this plugin.
  */
@@ -36,7 +34,8 @@ function atto_cloze_strings_for_js() {
     $PAGE->requires->strings_for_js(array( 'multichoice', 'numerical', 'shortanswer' ), 'mod_quiz' );
     $PAGE->requires->strings_for_js(array( 'addmoreanswerblanks', 'tolerance' ), 'qtype_calculated' );
     $PAGE->requires->strings_for_js(array( 'add', 'cancel', 'delete',
-            'duplicate', 'down', 'grade', 'previous', 'up' ), 'core' );
+            'duplicate', 'down', 'previous', 'up' ), 'core' );
+    $PAGE->requires->strings_for_js(array( 'grade' ), 'grades' );
 }
 
 /**